local/adapted/multimodal/requirements.txt
gtts==2.3.1
transformers==4.42.4
torch==1.13.1
flask==2.3.2
python-pptx==0.6.19
moviepy==1.0.3
numpy==1.21.2
sqlalchemy
moodleapi
requests